pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zabbix-mcp"
version = "0.0.1"
description = "MCP server for Zabbix management"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {file = "LICENSE"}
authors = [
    {name = "Mateusz Hajder"},
]
keywords = ["zabbix", "mcp", "network", "management", "server", "automation"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
    "Topic :: System :: Networking :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

dependencies = [
    "fastmcp>=2.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "zabbix-utils[async]>=2.0.4",
]

[project.optional-dependencies]
sentry = [
    "sentry-sdk>=2.43.0",
]

[dependency-groups]
dev = [
    "commitizen",
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pre-commit",
]

[project.scripts]
zabbix-mcp = "zabbix_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/zabbix_mcp"]

[project.urls]
Homepage = "https://github.com/mhajder/zabbix-mcp"
Repository = "https://github.com/mhajder/zabbix-mcp"
Documentation = "https://github.com/mhajder/zabbix-mcp#readme"
Issues = "https://github.com/mhajder/zabbix-mcp/issues"

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.11+.
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "N",  # pep8-naming
    "S",  # flake8-bandit
    "T20", # flake8-print
    "SIM", # flake8-simplify
    "LOG", # flake8-logging
    "RUF", # Ruff-specific rules
]

ignore = [
    "E501",  # Line too long, handled by formatter
    "S104",  # Possible binding to all interface
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106", "S107"]  # Allow assert and hardcoded passwords in tests

[tool.ruff.lint.isort]
known-first-party = ["zabbix_mcp"]
force-single-line = true

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
